+++ src/app/app.component.html
@@ -1,8 +1,8 @@
-<div class="box" [shContextMenuTriggerFor]="menu" [shMenuData]="items[0]">
+<div class="box" [shAnchorFor]="menu" [shMenuData]="items[0]">
   <h2>{{title}}</h2>
 </div>
 
-<div class="box" [shContextMenuTriggerFor]="myMenu" [shMenuData]="items[1]">
+<div class="box" [shAnchorFor]="myMenu" [shMenuData]="items[1]">
   <h2>{{title}}</h2>
 </div>
