+++ src/lib/src/sh-context-menu-item.directive.ts
@@ -1,0 +1,9 @@
+import {Directive, TemplateRef} from '@angular/core';
+
+@Directive({
+  selector: '[shContextMenuItem]'
+})
+export class ShContextMenuItemDirective {
+  constructor(public template: TemplateRef<any>) {
+  }
+}
